Players moving and armed
Players can now move around and kill each other. Once everyone dies there is nothing really to do.

Changes:
- Created a proper turn method for Player class
- In-game console now has slightly smaller text that is also now left-aligned
- Removed unused show players checkbox
- Controls are hidden if the corresponding checkbox is unchecked
- Treasure generation coordinates are now stored so that future features are not placed on it
- Player dots now have an id above them
- Added "wallToString" - a method used to create descriptions for wall types. Used to create console messages.

Most gameplay elements, such as treasure, wormholes, exits, etc. have not been implemented still, so there is not much gameplay value. But it can be played.

In the next few commits, I hope to start on AI things. I need to figure out how to store things in the AI's memory as well as revamp how the map is drawn to reflect this. This also applies to the currently unused canvases of the other players.